BOAC-1830 Break search out into a dedicated API, including notes
