+++ src/api/student.ts
@@ -1,6 +1,5 @@
 import axios from 'axios';
 import store from '@/store';
-import { event } from 'vue-analytics';
 
 export function dismissStudentAlert(alertId: string) {
   let apiBaseUrl = store.getters['context/apiBaseUrl'];
@@ -13,32 +12,5 @@ export function getStudent(uid: string) {
   let apiBaseUrl = store.getters['context/apiBaseUrl'];
   return axios
     .get(`${apiBaseUrl}/api/student/${uid}`)
-    .then(response => response.data, () => null);
-}
-
-export function search(
-  phrase: string,
-  includeCourses: boolean,
-  isInactiveAsc: boolean,
-  isInactiveCoe: boolean,
-  orderBy: string,
-  offset: number,
-  limit: number
-) {
-  let apiBaseUrl = store.getters['context/apiBaseUrl'];
-  return axios
-    .post(`${apiBaseUrl}/api/students/search`, {
-      searchPhrase: phrase,
-      includeCourses: includeCourses,
-      isInactiveAsc: isInactiveAsc,
-      isInactiveCoe: isInactiveCoe,
-      orderBy: orderBy || 'first_name',
-      offset: offset || 0,
-      limit: limit || 50
-    })
-    .then(response => {
-      event('Students', 'search', phrase, store.getters['user/user'].uid);
-      return response;
-    })
     .then(response => response.data, () => null);
 }
