+++ tests/test_api/test_student_controller.py
@@ -221,10 +221,10 @@ class TestAthleticsStudyCenter:
         assert inactive_student['athleticsProfile']['statusAsc'] == 'Trouble'
 
 
-class TestSearch:
-    """Student Search API."""
+class TestStudentResultsForFilter:
+    """Student API results for filtered cohort criteria."""
 
-    sample_search = {
+    sample_filter = {
         'gpaRanges': ['numrange(3, 3.5, \'[)\')', 'numrange(3.5, 4, \'[]\')'],
         'groupCodes': ['MFB-DB', 'MFB-DL'],
         'levels': ['Junior', 'Senior'],
@@ -241,183 +241,15 @@ class TestSearch:
         'limit': 50,
     }
 
-    asc_search = json.dumps(sample_search)
+    asc_filter = json.dumps(sample_filter)
 
-    sample_search['groupCodes'] = []
-    sample_search['gpaRanges'] = []
-    sample_search['levels'] = []
-    coe_search = json.dumps(sample_search)
-
-    def test_search_not_authenticated(self, client):
-        """Search is not available to the world."""
-        response = client.post('/api/students/search')
-        assert response.status_code == 401
-
-    def test_unauthorized_request_for_athletic_study_center_data(self, client, fake_auth):
-        """In order to access intensive_cohort, inactive status, etc. the user must be either ASC or Admin."""
-        fake_auth.login('1022796')
-        args = {'searchPhrase': 'John', 'isInactiveAsc': False}
-        response = client.post('/api/students/search', data=json.dumps(args), content_type='application/json')
-        assert response.status_code == 403
-
-    def test_search_with_missing_input(self, client, fake_auth):
-        """Search is nothing without input."""
-        fake_auth.login('2040')
-        response = client.post('/api/students/search', data=json.dumps({'searchPhrase': ' \t  '}), content_type='application/json')
-        assert response.status_code == 400
-
-    def test_search_by_sid_snippet(self, client, fake_auth, asc_inactive_students):
-        """Search by snippet of SID."""
-        def _search_students_as_user(uid, sid_snippet):
-            fake_auth.login(uid)
-            response = client.post(
-                '/api/students/search',
-                data=json.dumps({'searchPhrase': sid_snippet}),
-                content_type='application/json',
-            )
-            assert response.status_code == 200
-            return response.json['students'], response.json['totalStudentCount']
-
-        sid_snippet = '89012'
-        # Admin user performs search
-        students, total_student_count = _search_students_as_user('2040', sid_snippet)
-        assert len(students) == total_student_count == 2
-        assert _get_common_sids(asc_inactive_students, students)
-        # ASC advisor performs the same search
-        students, total_student_count = _search_students_as_user('1081940', sid_snippet)
-        assert len(students) == total_student_count == 1
-        assert not _get_common_sids(asc_inactive_students, students)
-
-    def test_alerts_in_search_results(self, client, create_alerts, fake_auth):
-        """Search results include alert counts."""
-        fake_auth.login('2040')
-        response = client.post('/api/students/search', data=json.dumps({'searchPhrase': 'davies'}), content_type='application/json')
-        assert response.status_code == 200
-        assert response.json['students'][0]['alertCount'] == 3
-
-    def test_summary_profiles_in_search_results(self, client, fake_auth):
-        fake_auth.login('2040')
-        response = client.post('/api/students/search', data=json.dumps({'searchPhrase': 'davies'}), content_type='application/json')
-        assert response.json['students'][0]['cumulativeGPA'] == 3.8
-        assert response.json['students'][0]['cumulativeUnits'] == 101.3
-        assert response.json['students'][0]['expectedGraduationTerm']['name'] == 'Fall 2019'
-        assert response.json['students'][0]['level'] == 'Junior'
-        assert response.json['students'][0]['termGpa'][0]['gpa'] == 2.9
-
-    def test_search_by_name_snippet(self, client, fake_auth):
-        """Search by snippet of name."""
-        fake_auth.login('2040')
-        response = client.post('/api/students/search', data=json.dumps({'searchPhrase': 'dav'}), content_type='application/json')
-        assert response.status_code == 200
-        students = response.json['students']
-        assert len(students) == response.json['totalStudentCount'] == 3
-        assert ['Crossman', 'Davies', 'Doolittle'] == [s['lastName'] for s in students]
-
-    def test_search_by_full_name_snippet(self, client, fake_auth):
-        """Search by snippet of full name."""
-        fake_auth.login('2040')
-        permutations = ['david c', 'john  david cro', 'john    cross', ' crossman, j ']
-        for phrase in permutations:
-            response = client.post(
-                '/api/students/search',
-                data=json.dumps({'searchPhrase': phrase}),
-                content_type='application/json',
-            )
-            message_if_fail = f'Unexpected result(s) when search phrase={phrase}'
-            assert response.status_code == 200, message_if_fail
-            students = response.json['students']
-            assert len(students) == response.json['totalStudentCount'] == 1, message_if_fail
-            assert students[0]['lastName'] == 'Crossman', message_if_fail
-
-    def test_search_by_name_asc_limited(self, asc_advisor, client):
-        """An ASC name search finds active ASC Pauls."""
-        response = client.post('/api/students/search', data='{"searchPhrase": "paul"}', content_type='application/json')
-        students = response.json['students']
-        assert len(students) == 2
-        assert next(s for s in students if s['name'] == 'Paul Kerschen')
-        assert next(s for s in students if s['name'] == 'Paul Farestveit')
-
-    def test_search_by_name_coe_limited(self, coe_advisor, client):
-        """A COE name search finds active COE Pauls."""
-        response = client.post('/api/students/search', data='{"searchPhrase": "paul"}', content_type='application/json')
-        students = response.json['students']
-        assert len(students) == 1
-        assert next(s for s in students if s['name'] == 'Paul Farestveit')
-
-    def test_search_by_name_admin_unlimited(self, admin_login, client):
-        """An admin name search finds all Pauls."""
-        response = client.post('/api/students/search', data='{"searchPhrase": "paul"}', content_type='application/json')
-        students = response.json['students']
-        assert len(students) == 3
-        assert next(s for s in students if s['name'] == 'Paul Kerschen')
-        assert next(s for s in students if s['name'] == 'Paul Farestveit')
-        assert next(s for s in students if s['name'] == 'Wolfgang Pauli')
-
-    def test_search_by_name_excludes_courses_unless_requested(self, coe_advisor, client):
-        response = client.post('/api/students/search', data='{"searchPhrase": "da"}', content_type='application/json')
-        assert 'courses' not in response.json
-        assert 'totalCourseCount' not in response.json
-
-    def test_search_by_name_includes_courses_if_requested(self, coe_advisor, client):
-        """A name search returns matching courses if any."""
-        response = client.post(
-            '/api/students/search',
-            data=json.dumps({'searchPhrase': 'paul', 'includeCourses': True}),
-            content_type='application/json',
-        )
-        assert response.json['courses'] == []
-        response = client.post(
-            '/api/students/search',
-            data=json.dumps({'searchPhrase': 'da', 'includeCourses': True}),
-            content_type='application/json',
-        )
-        students = response.json['students']
-        assert len(students) == 1
-        assert students[0]['name'] == 'Deborah Davies'
-        courses = response.json['courses']
-        assert len(courses) == 1
-        assert response.json['totalCourseCount'] == 1
-        assert courses[0] == {
-            'termId': '2178',
-            'sectionId': '21057',
-            'courseName': 'DANISH 1A',
-            'courseTitle': 'Beginning Danish',
-            'instructionFormat': 'LEC',
-            'sectionNum': '001',
-            'instructors': 'Karen Blixen',
-        }
-
-    def test_search_by_name_normalizes_queries(self, coe_advisor, client):
-        queries = ['MATH 16A', 'Math 16 A', 'math  16-a']
-        for query in queries:
-            response = client.post(
-                '/api/students/search',
-                data=json.dumps({'searchPhrase': query, 'includeCourses': True}),
-                content_type='application/json',
-            )
-            courses = response.json['courses']
-            assert len(courses) == 2
-            assert response.json['totalCourseCount'] == 2
-            for course in courses:
-                assert course['courseName'] == 'MATH 16A'
-
-    def test_search_order_by_offset_limit(self, client, fake_auth):
-        """Search by snippet of name."""
-        fake_auth.login('2040')
-        args = {
-            'searchPhrase': 'dav',
-            'orderBy': 'major',
-            'offset': 1,
-            'limit': 1,
-        }
-        response = client.post('/api/students/search', data=json.dumps(args), content_type='application/json')
-        assert response.status_code == 200
-        assert response.json['totalStudentCount'] == 3
-        assert len(response.json['students']) == 1
-        assert 'Crossman' == response.json['students'][0]['lastName']
+    sample_filter['groupCodes'] = []
+    sample_filter['gpaRanges'] = []
+    sample_filter['levels'] = []
+    coe_filter = json.dumps(sample_filter)
 
     def test_get_students(self, asc_advisor, asc_inactive_students, client):
-        response = client.post('/api/students', data=self.asc_search, content_type='application/json')
+        response = client.post('/api/students', data=self.asc_filter, content_type='application/json')
         assert response.status_code == 200
         assert 'students' in response.json
         students = response.json['students']
@@ -427,7 +259,7 @@ class TestSearch:
         assert ['9933311', '242881'] == [student['uid'] for student in students]
 
     def test_get_students_includes_athletics_asc(self, asc_advisor, client):
-        response = client.post('/api/students', data=self.asc_search, content_type='application/json')
+        response = client.post('/api/students', data=self.asc_filter, content_type='application/json')
         students = response.json['students']
         group_codes_1133399 = [a['groupCode'] for a in students[0]['athleticsProfile']['athletics']]
         assert len(group_codes_1133399) == 3
@@ -438,7 +270,7 @@ class TestSearch:
         assert group_codes_242881 == ['MFB-DL']
 
     def test_coe_unauthorized_request_for_asc_data(self, coe_advisor, client):
-        response = client.post('/api/students', data=self.coe_search, content_type='application/json')
+        response = client.post('/api/students', data=self.coe_filter, content_type='application/json')
         assert 403 == response.status_code
 
     def test_get_active_asc_students(self, asc_advisor, asc_inactive_students, client):
@@ -451,7 +283,7 @@ class TestSearch:
         assert not len(students)
 
     def test_get_inactive_asc_students(self, asc_advisor, asc_inactive_students, client):
-        """An ASC cohort search finds ASC sophomores."""
+        """ASC cohort results include ASC sophomores."""
         args = {
             'levels': ['Sophomore'],
             'isInactiveAsc': True,
@@ -464,14 +296,14 @@ class TestSearch:
         assert next(s for s in students if s['name'] == 'Siegfried Schlemiel')
 
     def test_get_students_coe_limited(self, coe_advisor, client):
-        """A COE cohort search finds active COE sophomores."""
+        """COE cohort results include active COE sophomores."""
         response = client.post('/api/students', data='{"levels": ["Sophomore"]}', content_type='application/json')
         students = response.json['students']
         assert len(students) == 1
         assert next(s for s in students if s['name'] == 'Nora Stanton Barney')
 
     def test_get_students_admin_unlimited(self, admin_login, client):
-        """An admin cohort search finds all sophomores."""
+        """Admin cohort results include all sophomores."""
         response = client.post('/api/students', data='{"levels": ["Sophomore"]}', content_type='application/json')
         students = response.json['students']
         assert len(students) == 3
